ieee802154: adf702x: add sport mode support
It is possible to bypass all of the packet management features of the
ADF702x and use the sport interface for transmit and receive data.
The sport interface is a high speed synchronous serial interface
allowing direct interfacing to processors and DSPs.

In sport mode, the interrupts from INTERRUPT_SOURCE_1 are all available.
However, only INTERRUPT_PREAMBLE_DETECT and INTERRUPT_SYNC_DETECT are
available from INTERRUPT_SOURCE_0.

A second interrupt pin is provided on GP4, which gives a dedicated sport
mode interrupt on either preamble or sync word detection. For more
details, see the Sport Mode section.

In the current implementation, the second IRQ pin is left disabled.

Signed-off-by: Liam Beguin <liambeguin@gmail.com>